requirements.txt
playwright
pandas
streamlit
tqdm
Pillow